Synced namespaces, and moved all item type swaps to their own file and folder
