+++ Saturn.Backend/Data/SwapOptions/Pickaxes/GumBrawlerSwap.cs
@@ -2,9 +2,8 @@
 using Saturn.Backend.Data.Enums;
 using Saturn.Backend.Data.Models.Items;
 using Saturn.Backend.Data.Services;
-using Saturn.Backend.Data.Utils.Swaps;
 
-namespace Saturn.Backend.Data.SwapOptions.Pickaxe;
+namespace Saturn.Backend.Data.SwapOptions.Pickaxes;
 
 internal sealed class GumBrawlerSwap : PickaxeSwap
 {
