+++ Saturn.Backend/Data/SwapOptions/Pickaxes/MayhemScytheSwap.cs
@@ -1,9 +1,8 @@
 ﻿using System.Collections.Generic;
 using Saturn.Backend.Data.Enums;
 using Saturn.Backend.Data.Models.Items;
-using Saturn.Backend.Data.Utils.Swaps;
 
-namespace Saturn.Backend.Data.SwapOptions.Pickaxe;
+namespace Saturn.Backend.Data.SwapOptions.Pickaxes;
 
 internal sealed class MayhemScytheSwap : PickaxeSwap
 {
