+++ Saturn.Backend/Data/SwapOptions/Pickaxes/ShadowSlicerSwap.cs
@@ -3,9 +3,8 @@ using System.Net.Sockets;
 using Saturn.Backend.Data.Enums;
 using Saturn.Backend.Data.Models.Items;
 using Saturn.Backend.Data.Services;
-using Saturn.Backend.Data.Utils.Swaps;
 
-namespace Saturn.Backend.Data.SwapOptions.Pickaxe;
+namespace Saturn.Backend.Data.SwapOptions.Pickaxes;
 
 internal sealed class ShadowSlicerSwap : PickaxeSwap
 {
