+++ Saturn.Backend/Data/SwapOptions/Skins/SkinSwap.cs
@@ -1,0 +1,15 @@
+﻿using Saturn.Backend.Data.Models.Items;
+using Saturn.Backend.Data.Utils.Swaps;
+
+namespace Saturn.Backend.Data.SwapOptions.Skins;
+
+internal abstract class SkinSwap : AbstractSwap
+{
+    protected SkinSwap(string name, string rarity, string icon, MeshDefaultModel swapModel)
+        : base(name, rarity, icon)
+    {
+        SwapModel = swapModel;
+    }
+
+    public MeshDefaultModel SwapModel { get; }
+}
